requirements.txt
opencv-python
pytesseract
python-dotenv